+++ feature/home/src/main/java/com/dida/home/adapter/CollectionAdapter.kt
@@ -32,7 +32,7 @@ class CollectionAdapter(
         holder.bind(getItem(position))
     }
 
-    override fun getItemId(position: Int): Long = getItem(position).userId.toLong() * -1
+    override fun getItemId(position: Int): Long = getItem(position).userId * -1
 
     override fun getItemViewType(position: Int): Int = R.layout.holder_collection
