+++ feature/nft-detail/src/main/java/com/dida/nft_detail/DetailNftNavigationAction.kt
@@ -5,7 +5,7 @@ sealed class DetailNftNavigationAction {
     class NavigateToItemCommunity(val postId: Long): DetailNftNavigationAction()
     object NavigateToCreateCommunity: DetailNftNavigationAction()
     object NavigateToHome : DetailNftNavigationAction()
-    class NavigateToUserProfile(val userId: Int): DetailNftNavigationAction()
+    class NavigateToUserProfile(val userId: Long): DetailNftNavigationAction()
     object NavigateToBack : DetailNftNavigationAction()
     object NavigateToSell : DetailNftNavigationAction()
 }
